Make bodies traversable. Make trailers different.
This change makes it more awkward to silently treat a trailer chunk
like a body chunk.  This is attractive because it forces the backend
to consider the difference and perhaps log inappropriate use.  This is
bad because trailers are rare and only Netty supports them.

Alternatively, we might make the HttpChunk a Traversable[Byte].  This
would hide the difference for those who don't care, but give rise to
awkward cases like (Body ++ Trailer => Body).