My take on async servlet streams
This implementation should be rather efficient and only block in the rare event that the stream is attempted to be read by two threads who both enter the while(isReady) loop of doRead at the same time. It is clearly an error to be reading raw bytes from the same stream in different threads except in the situation that each byte is a self contained message. If that is the situation, sadface.