Add a ResponseException type that can be converted to a Response
This would provide a way to allow failures to have a default representation that the backend can render if the service doesn't handle it. This will work well for things like decoders where the user doesn't typically want to deal with a parsing exception, but the client should be notified that there was a problem related to it.