Remove ReplyException. http4s/http4s#43 http4s/http4s#96
This punts on making errors explicit in the Client: those still get
thrown as a ParseException.

There is a good argument to be made that the apply method added to
EntityDecoder is server DSL syntax, not functionality of the
EntityDecoder.  It also kind of sucks to use an EntityDecoder on the
server without it.