Don't accidentally load non-blocking servlet API. http4s/http4s#229
The check in Http4sServlet comes too late: NonBlockingServletIo is the
default, and it already tries to load classes that aren't there.  It's
fine in the xsbt-web-plugin with Jetty 9, but Tomcat 7 exposes the
problem.

Make the ServletContextOps default to the proper servlet I/O version.
Constructing a servlet by hand will now default to blocking I/O.  When
we can break binary compatibility, we will reinstate the option to
select based on the server.