Remove our own `DateTime` for java.time.Instant
closes http4s/http4s#476

This also addresses a bug in our header parser where the day of the month was required to be 2 digits instead of 1  *or* 2 digits. The ABNF is:
1*2DIGIT
in RFC5322, RFC1123, and RFC850
http://tools.ietf.org/html/rfc5322http4s/http4s#section-3.3

However, in RFC7231 it is specified as
day = 2DIGIT

Instant's are formatted according to RFC7231 format for HTTP usage.