Fix infinite loop in nonblocking ServletIo
Problem: If the state gets to `Ready` and the stream is at EOF, the process will receive an infinite stream of `ByteVector.empty` because the process is being run on the servlet thread and therefor the listener will not have toe opportunity to call `onAllDataRead()` on the listener.

Solution: With an InputStream EOF is signaled as a read length of -1 bytes. The read method now looks for this condition and terminates the stream itself.

Additional changes:
* Fixed the Tomcat bits of the build to actually use the right main class in SBT-revolver
* Added a MiMa exception for a method that shouldn't actually be public but MiMa thinks it is.

Future concerns: the Tomcat example is broken when using SBT-revolver: it fails to load the metrics bits because it cant find the class for some reason.