Let the the ReadListener be the first one to send bytes
Problem: The previous commit was causing Jetty to murder itself for some reason. It appears that when using async mode, it wants to be in charge of performing the first read through its call to `onDataAvailable` in the `ReadListener`.

Solution: This commit makes the input Process consist of an initialization step that sets the listener and waits for bytes and subsequent 'normal read' steps that check the state of things.