Change Uri.Path to be own type
This changes Uri.Path to be an actual class to encapsulate Path segments.
We have two boolean flags here. These can be removed by transforming
the `Path` type to an own `ADT` hierarchy.

The main motivation for this change is to represent paths as something more than a String.

I have used `scala.collection.immutable.Vector` instead of `List` as `Vector` allows O(1)
appends.

This change is binary incompatible.

Typeclasses:
* Eq
* Semigroup
* Adding a monoid is possible, but there are two zero cases. Changing the Path type to an ADT will resolve that.

We have a few convenince functions here:

* startWith to be able to use in the Routers
* indexOf and splitAt to allow us to replace the caret stuff.
* normalize: drops empty segments
* toAbsolute and toRelative
* merge: using RFC3986 rules
* concat

Unrelated, add `out` to .gitignore to be able to use bloop from intellij

Representing Encoded Strings as just Strings are not really useful, so we have added a Segment class to have encoded values in.

* Use uri and path string interpolation in tests

Request changes:
* scriptName and pathInfo changed to be Uri.Path